requirements.txt
asttokens==2.0.8
backcall==0.2.0
backports.functools-lru-cache==1.6.4
cassandra-driver==3.25.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.38.0
geomet==0.2.1.post1
gunicorn==20.1.0
idna==3.4
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==5.0.0
ipykernel==6.16.1
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.3
jupyter_core==4.11.1
kiwisolver==1.4.4
kneed==0.8.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
prompt-toolkit==3.0.31
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
pyzmq==23.2.0
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.1
setuptools==63.4.1
six==1.16.0
stack-data==0.5.1
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
xgboost==1.7.1
zipp==3.10.0
